// src/app/core/models/employee.interface.ts
export interface IEmployee {
   id_Empleado?: number;
   nombre: string;
   apellido: string;
   fecha_Nacimiento: string;
   direccion: string;
   telefono: string;
   email: string;
   fecha_Contratacion: string;
   puesto: string;
   vacaciones: boolean;
   licencia: boolean;
   salario: number;
   estado: boolean;
   departamento_Id: number;
 }

 